backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["domin <domin.dober@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
elevenlabs = "^2.6.0"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
